Pandas gives a warning when creating a series with an empty (None) variable; when using pyOptimalEstimation and we do not provide xtruth, this warning is constantly on screen. We add here then the dtype to the creation of the xtruth series (np.float64)

+++ pyOptimalEstimation/pyOEcore.py
@@ -216,7 +216,7 @@ class optimalEstimation(object):
         self.y_n = len(self.y_vars)
         self.forward = forward
         self.userJacobian = userJacobian
-        self.x_truth = pd.Series(x_truth, index=self.x_vars)
+        self.x_truth = pd.Series(x_truth, index=self.x_vars, dtype=np.float64)
 
         # We want to save at least the name because the forward function
         # is removed for saving
